Execute balanced ANOVA for multiple indexes

This function fits an analysis of variance model (ANOVA) for various indexes. Input must be a table with indexes as columns and samples as rows. Further columns should include metadata.

BalancedAnova(data, numberOfIndexes, formula, ...)

Arguments

data
a data frame, columns corresponding to indexes and rows corresponding to samples. Further columns should be included with metadata. This is used in argument Formula

numberOfIndexes
Integer corresponding to the number of indexes to analyze. This will be take as column numbers by the function.

formula
Metadata group name. This will group samples according to a metadata column and specify the model.

...
Further arguments passed to aov.

Value

Returns a list of two elements. First element is a data frame, containing the summary statistics from aov on different indexes (indexes names are shown in the IndexColumn). Second element is a list, composed of as many list as indexes were analyzed. These contain the complete results from aov.

Examples

balanced_anova<- BalancedAnova(alpha_diversity_table, numberOfIndexes = 4, formula = "location")