Dunn's Test for multiple variables

This function make use of dunn.test function to compute Dunn's test of multiple comparisons along groups (using rank sums) and looping it along multiple variables of interest (e.g., alpha diversity indexes).

dunnT(data, numberOfIndexes, formula, dunn.options, method = "BH")

Arguments

data
a data frame, columns corresponding to indexes and rows corresponding to samples. Further columns should be included with metadata. This is used in argument formula. Further details to be found in dunn.test.

numberOfIndexes
Integer corresponding to the number of indexes to analyze. This will be taken as column numbers by the function.

formula
Metadata group name. This will group samples according to a metadata column (corresponding to g argument in dunn.test, representing grouping vector or factor).

dunn.options
Further arguments to be passed to dunn.test

method
Method selected to adjust the p-values for multiple comparisons. See dunn.test for more detailes about methods and abbreviations. Default is set to Benjamini-Hochberg adjustment ("BH")

Value

Returns a data frame with dunn.test results for all pairwise comparisons, performed on each variable (determined by numberOfIndexes).For further details on parameters, check dunn.test.

Examples

dunn_location <- dunnT(alpha_diversity_table,4,"location")

#>   Kruskal-Wallis rank sum test
#> 
#> data: x and group
#> Kruskal-Wallis chi-squared = 1.3386, df = 2, p-value = 0.51
#> 
#> 
#>                            Comparison of x by group                            
#>                              (Benjamini-Hochberg)                              
#> Col Mean-|
#> Row Mean |   location   location
#> ---------+----------------------
#> location |  -0.866960
#>          |     0.2895
#>          |
#> location |   0.230009   1.096969
#>          |     0.4090     0.4090
#> 
#> alpha = 0.05
#> Reject Ho if p <= alpha/2
#>   Kruskal-Wallis rank sum test
#> 
#> data: x and group
#> Kruskal-Wallis chi-squared = 2.405, df = 2, p-value = 0.3
#> 
#> 
#>                            Comparison of x by group                            
#>                              (Benjamini-Hochberg)                              
#> Col Mean-|
#> Row Mean |   location   location
#> ---------+----------------------
#> location |  -0.494974
#>          |     0.3103
#>          |
#> location |   1.025304   1.520279
#>          |     0.2289     0.1927
#> 
#> alpha = 0.05
#> Reject Ho if p <= alpha/2
#>   Kruskal-Wallis rank sum test
#> 
#> data: x and group
#> Kruskal-Wallis chi-squared = 4.5, df = 2, p-value = 0.11
#> 
#> 
#>                            Comparison of x by group                            
#>                              (Benjamini-Hochberg)                              
#> Col Mean-|
#> Row Mean |   location   location
#> ---------+----------------------
#> location |  -1.060660
#>          |     0.1444
#>          |
#> location |   1.060660   2.121320
#>          |     0.2166     0.0508
#> 
#> alpha = 0.05
#> Reject Ho if p <= alpha/2
#>   Kruskal-Wallis rank sum test
#> 
#> data: x and group
#> Kruskal-Wallis chi-squared = 1.5, df = 2, p-value = 0.47
#> 
#> 
#>                            Comparison of x by group                            
#>                              (Benjamini-Hochberg)                              
#> Col Mean-|
#> Row Mean |   location   location
#> ---------+----------------------
#> location |   0.000000
#>          |     0.5000
#>          |
#> location |   1.060660   1.060660
#>          |     0.2166     0.4333
#> 
#> alpha = 0.05
#> Reject Ho if p <= alpha/2