Levene's test for multiple variables

Test homogeneity of variance across groups, looping across multiple variables (e.g. alpha diversity indexes).

levene.test.alpha(data, numberOfIndexes, formula, ...)

Arguments

data
a data frame, columns corresponding to variables (e.g. alpha diversity indexes) and rows corresponding to samples. Further columns should be included with metadata. This is used in argument Formula.

numberOfIndexes
Integer corresponding to the number of indexes to analyze. This will be take as column numbers by the function.

formula
Metadata group name. This will group samples according to a metadata column and specify the model.

...
Further arguments to be passed to leveneTest.

Value

returns a data frame with results from leveneTest from package car for each variable (determined by numberOfIndexes).Further details can be found in leveneTest

Examples

levene_location <- Levene.test.alpha(alpha_diversity_table, 4, "location")

#> Error in Levene.test.alpha(alpha_diversity_table, 4, "location"): no se pudo encontrar la función "Levene.test.alpha"